+++ test/bigtest/network/configs/vouchers.js
@@ -1,6 +1,7 @@
 import {
   createGetAll,
   createGetById,
+  createPut,
 } from '@folio/stripes-acq-components/test/bigtest/network/configs';
 
 import { VOUCHERS_API } from '../../../../src/common/constants/api';
@@ -10,6 +11,7 @@ const SCHEMA_NAME = 'vouchers';
 const configVouchers = server => {
   server.get(VOUCHERS_API, createGetAll(SCHEMA_NAME));
   server.get(`${VOUCHERS_API}/:id`, createGetById(SCHEMA_NAME));
+  server.get(`${VOUCHERS_API}/:id`, createPut(SCHEMA_NAME));
 };
 
 export default configVouchers;